pyproject.toml
[project]
name = "gemini-ocr-cli"
version = "0.2.0"
description = "CLI tool for OCR processing using Google Gemini's vision capabilities"
authors = [
    {name = "Ruben Fernandez-Fuertes", email = "fernandezfuertesruben@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["ocr", "gemini", "google", "cli", "pdf", "document-processing", "vision"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: General",
]

dependencies = [
    "google-genai>=1.0.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
    "pymupdf>=1.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/r-uben/gemini-ocr-cli"
Repository = "https://github.com/r-uben/gemini-ocr-cli"
Issues = "https://github.com/r-uben/gemini-ocr-cli/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.8.0",
    "mypy>=1.0.0",
]

[project.scripts]
gemini-ocr = "gemini_ocr.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["gemini_ocr"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
markers = [
    "unit: marks tests as unit tests (fast, no API)",
    "integration: marks tests requiring GEMINI_API_KEY",
]
